pacrunner: do not autostart service
pacrunner is supposed to be autostarted by clients, not by managers
like NM, so that it runs only when necessary. When a client issues a
request the service is started and waits some time before responding
so that managers have the chance to notice its presence on the bus and
push the proxy configuration.

Thus, pass the NO_AUTO_START flag to GDBus calls; also, avoid any
D-Bus call if we know pacrunner is not running.